+++ src/Monster.java
@@ -18,24 +18,24 @@ public class Monster {
 
 
     //Properties (Constant)
-    private static int healthMax;
-    private static int manaMax;
+    private static double healthMax;
+    private static double manaMax;
     private double speedAttack;     //Znalezc sensowne zastosowanie
     private int lvl;        //Znalezc sensowne zastosowanie
     private static String nameOfTheMonster;
-    private static int expMin;
-    private static int expMax;
-    private static int coinDropMin;
-    private static int coinDropMax;
-    private static int damageMin;
-    private static int damageMax;
-    private static int minMagicDmg;
-    private static int maxMagicDmg;
+    private static double expMin;
+    private static double expMax;
+    private static double coinDropMin;
+    private static double coinDropMax;
+    private static double damageMin;
+    private static double damageMax;
+    private static double minMagicDmg;
+    private static double maxMagicDmg;
 
 
     //Variables
-    private static int health;
-    private static int mana;       //Znalezc sensowne zastosowanie
+    private static double health;
+    private static double mana;       //Znalezc sensowne zastosowanie
     private int healthPotions;
 
 
@@ -44,11 +44,11 @@ public class Monster {
         return this.healthPotions;
     }
 
-    public static int getHealth() {
+    public static double getHealth() {
         return health;
     }
 
-    public static int getHealthMax() {
+    public static double getHealthMax() {
         return healthMax;
     }
 
@@ -56,11 +56,11 @@ public class Monster {
         return (health + "/" + healthMax);
     }
 
-    public static int getMana() {
+    public static double getMana() {
         return mana;
     }
 
-    public static int getManaMax() {
+    public static double getManaMax() {
         return manaMax;
     }
 
@@ -85,19 +85,19 @@ public class Monster {
         return lvl;
     }
 
-    public static int getMinMagicDmg() {
+    public static double getMinMagicDmg() {
         return minMagicDmg;
     }
 
-    public static int getMaxMagicDmg() {
+    public static double getMaxMagicDmg() {
         return maxMagicDmg;
     }
 
-    public static int getDamageMin() {
+    public static double getDamageMin() {
         return damageMin;
     }
 
-    public static int getDamageMax() {
+    public static double getDamageMax() {
         return damageMax;
     }
 
@@ -121,12 +121,12 @@ public class Monster {
         this.coinDropMax = max;
     }
 
-    public static void setHealth(int current, int max) {
+    public static void setHealth(double current, double max) {
         health = current;
         healthMax = max;
     }
 
-    public static void setMana(int current, int max) {
+    public static void setMana(double current, double max) {
         mana = current;
         manaMax = max;
     }
@@ -137,7 +137,7 @@ public class Monster {
 
 
     //Monster constructor
-    public Monster(int health, int healthMax, int mana, int manaMax, double speedAttack, int lvl, String nameOfTheMonster, int expMin, int expMax, int coinDropMin, int coinDropMax, int damageMin, int damageMax, int minMagicDmg, int maxMagicDmg) {
+    public Monster(double health, double healthMax, double mana, double manaMax, double speedAttack, int lvl, String nameOfTheMonster, double expMin, double expMax, double coinDropMin, double coinDropMax, double damageMin, double damageMax, double minMagicDmg, double maxMagicDmg) {
         this.health = health;
         this.healthMax = healthMax;
         this.mana = mana;
@@ -162,7 +162,7 @@ public class Monster {
         current.healthPotions = Random.RInt(HEALTH_POTIONS_MIN, HEALTH_POTIONS_MAX);
     }
 
-    public static boolean takeDamage(int damage) {
+    public static boolean takeDamage(double damage) {
         health = health - damage;
         System.out.println(getNameOfTheMonster() + " przyjal obrazenia w wysokosci: " + damage + ", ma teraz: " + health + " punktow zycia");
         if (health <= 0) {
@@ -172,7 +172,7 @@ public class Monster {
         return false;
     }
 
-    public static void dealDamage(int damage) {
+    public static void dealDamage(double damage) {
         System.out.println(nameOfTheMonster + " zadal Ci: " + damage + " punktow obrazen");
         /* Dodac przyjmowanie obrazen - coś w stylu playerHealth.takeDamage(damage) */
     }
@@ -181,8 +181,8 @@ public class Monster {
         /* złapać pomysł do napisana */
 
         //Rewards
-        int coin = Random.RInt(coinDropMin, coinDropMax);
-        int exp = Random.RInt(expMin, expMax);
+        double coin = Random.RInt(coinDropMin, coinDropMax);
+        double exp = Random.RInt(expMin, expMax);
 
         System.out.println("Zajebałeś stworka, dostajesz: " + coin + " monet i " + exp + " punktów doświadczenia!");
 
@@ -204,8 +204,8 @@ public class Monster {
 
     public void magicalAttack(int mana){
         Magic Magic = new Magic();
-        int typeOfAttack;
-        int damage;
+        double typeOfAttack;
+        double damage;
 
         if (mana >= 20 && mana < 30){
             damage = Magic.useFireBall(getLvl(), getMinMagicDmg(), getMaxMagicDmg());
@@ -244,7 +244,7 @@ public class Monster {
     }
 
     public void physicalAttack(int damageMin, int damageMax){
-        int damage = Random.RInt(damageMin, damageMax);
+        double damage = Random.RInt(damageMin, damageMax);
         dealDamage(damage);
     }
